Add files via upload
Path Sum: You are given a binary tree in which each node is an integer value(which might be positive or negative). 
Design an algorithm to count the number of paths that sum to a given value. The path does not need to start or end at the root or a leaf, but must go downward. (traveling from parent to node.)